requirements.txt
asgiref==3.8.1
Django==5.1.2
django-cors-headers==4.6.0
django-filter==24.3
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
gunicorn==23.0.0
Markdown==3.7
mysqlclient==2.2.5
packaging==24.2
phonenumbers==8.13.47
python-dotenv==1.0.1
sqlparse==0.5.1
tzdata==2024.2
